Improving behavior for moving to next and previous question.
Locally saving the starred questions
