--- Leben In Deutshland/Questions/OptionView.swift
@@ -1,24 +1,0 @@
-//
-//  OptionView.swift
-//  Leben In Deutshland
-//
-//  Created by Hassaan Ahmed on 16.03.25.
-//
-
-import SwiftUI
-
-//struct OptionView: View {
-//    let text: String
-//    let isSelected: Bool
-//    
-//    var body: some View {
-//        HStack {
-//            Image(systemName: isSelected ? "checkmark.square.fill" : "square")
-//            Text(text)
-//        }
-//    }
-//}
-
-#Preview {
-    OptionView(text: "This is some Question", isSelected: false, isCorrect: false, correctAnswer: "b")
-}
